// frontend/src/types/enums.ts
export {
    ModeTxt,
    ModeClass,
    FanTxt,
    FanClass,
    ComTxt,
    ComClass,
    StatusTxt,
    StatusClass,
    AlarmTxt,
    AlarmClass
}

enum ModeTxt {
    Cool = 'COOL',
    Heat = 'HEAT',
    Fan = 'FAN',
    Auto = 'AUTO'
}

enum ModeClass {
    Cool = 'cool',
    Heat = 'heat',
    Fan = 'fan',
    Auto = 'auto'
}

enum FanTxt {
    High = 'High',
    Medium = 'Medium',
    Low = 'Low',
    Auto = 'Auto'
}

enum FanClass {
    High = 'high',
    Medium = 'medium',
    Low = 'low',
    Auto = 'auto'
}

enum ComTxt {
    Off = 'OFF',
    On = 'ON',
    Auto = 'AUTO'
}

enum ComClass {
    Off = 'off',
    On = 'on',
    Auto = 'auto'
}

enum StatusTxt {
    Off = 'OFF',
    On = 'ON'
}

enum StatusClass {
    Off = 'off',
    On = 'on'
}

enum AlarmTxt {
    Ok = 'OK',
    Alarm = 'ALARM',
    Delay = 'OK*'
}

enum AlarmClass {
    Ok = 'ok',
    Alarm = 'alarm'
}